changed rate type from int to float

+++ Movie.h
@@ -25,7 +25,7 @@ class Movie
         std::string director;
         int id;
         std::vector<int> rates;
-        int rate;
+        float rate;
 };
 
 #endif
